refactor: clean up DTO classes and improve deputy expenses handling
- Remove unused DTO classes from response DTOs to reduce code duplication
- Move expense calculation methods from controller to Deputy model
- Add availableYears(), monthlyExpenses(), and expensesByCategory() methods to Deputy model
- Limit deputy search results to 10 entries for better performance
- Add toArray() method to LinkDTO for consistency
- Remove hardcoded mock labels from deputy show view
- Improve separation of concerns by moving business logic to model layer
