+++ app/DTOs/OpenData/DeputadoFrentesResponseDTO.php
@@ -13,22 +13,3 @@ class DeputadoFrentesResponseDTO
         public array $links = []
     ) {}
 }
-
-class FrenteDTO
-{
-    public function __construct(
-        public int $idFrente,
-        public string $titulo,
-        public string $uriFrente,
-        public string $situacao
-    ) {}
-}
-
-class LinkDTO
-{
-    public function __construct(
-        public string $href,
-        public string $rel,
-        public string $type
-    ) {}
-}
