+++ app/DTOs/OpenData/DeputadoOcupacoesResponseDTO.php
@@ -14,14 +14,3 @@ class DeputadoOcupacoesResponseDTO
     ) {
     }
 }
-
-class OcupacaoDTO
-{
-    public function __construct(
-        public string $titulo,
-        public string $entidade,
-        public string $inicio,
-        public string $fim
-    ) {
-    }
-}
