+++ app/DTOs/OpenData/DeputadoSituacaoDTO.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\DTOs\OpenData;
+
+class DeputadoSituacaoDTO
+{
+    public function __construct(
+        public int $id,
+        public string $descricao
+    ) {
+    }
+}
