+++ app/DTOs/OpenData/LegislaturaLideresResponseDTO.php
@@ -14,16 +14,3 @@ class LegislaturaLideresResponseDTO
     ) {
     }
 }
-
-class LiderDTO
-{
-    public function __construct(
-        public int $idDeputado,
-        public string $nomeDeputado,
-        public string $siglaPartido,
-        public string $siglaUf,
-        public string $tipo,
-        public string $uri
-    ) {
-    }
-}
