+++ app/DTOs/OpenData/LegislaturaMesasResponseDTO.php
@@ -14,16 +14,3 @@ class LegislaturaMesasResponseDTO
     ) {
     }
 }
-
-class MesaDTO
-{
-    public function __construct(
-        public int $idDeputado,
-        public string $nome,
-        public string $siglaPartido,
-        public string $siglaUf,
-        public string $titulo,
-        public string $uri
-    ) {
-    }
-}
