+++ app/DTOs/OpenData/ReferenciaDeputadoSituacaoResponseDTO.php
@@ -14,12 +14,3 @@ class ReferenciaDeputadoSituacaoResponseDTO
     ) {
     }
 }
-
-class DeputadoSituacaoDTO
-{
-    public function __construct(
-        public int $id,
-        public string $descricao
-    ) {
-    }
-}
