+++ app/DTOs/OpenData/SituacaoDeputadoDTO.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace App\DTOs\OpenData;
+
+class SituacaoDeputadoDTO
+{
+    public function __construct(
+        public int $id,
+        public string $descricao
+    ) {
+    }
+}
