+++ app/DTOs/OpenData/SituacoesDeputadoResponseDTO.php
@@ -14,12 +14,3 @@ class SituacoesDeputadoResponseDTO
     ) {
     }
 }
-
-class SituacaoDeputadoDTO
-{
-    public function __construct(
-        public int $id,
-        public string $descricao
-    ) {
-    }
-}
